[centos-*] Improve DRYness, correctness, and speed of CentOS templates.

+++ scripts/common/minimize.sh
@@ -1,12 +1,15 @@
 #!/bin/sh -eux
 
-# Whiteout the swap partition to reduce box size
-# Swap is disabled till reboot
 swapuuid="`/sbin/blkid -o value -l -s UUID -t TYPE=swap`";
-swappart="`readlink -f /dev/disk/by-uuid/$swapuuid`";
-/sbin/swapoff "$swappart";
-dd if=/dev/zero of="$swappart" bs=1M || echo "dd exit code $? is suppressed";
-/sbin/mkswap -U "$swapuuid" "$swappart";
+
+if [ "x${swapuuid}" != "x" ]; then
+    # Whiteout the swap partition to reduce box size
+    # Swap is disabled till reboot
+    swappart="`readlink -f /dev/disk/by-uuid/$swapuuid`";
+    /sbin/swapoff "$swappart";
+    dd if=/dev/zero of="$swappart" bs=1M || echo "dd exit code $? is suppressed";
+    /sbin/mkswap -U "$swapuuid" "$swappart";
+fi
 
 dd if=/dev/zero of=/EMPTY bs=1M || echo "dd exit code $? is suppressed";
 rm -f /EMPTY;
